# api/openapi/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ApiError": {
				"properties": {
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"details": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"error",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PaymentSourceType": {
				"type": "string",
				"enum": [
					"bank_account",
					"credit_card",
					"line_of_credit",
					"cash"
				]
			},
			"PaymentSource": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/PaymentSourceType"
					},
					"balance": {
						"type": "number",
						"format": "double"
					},
					"is_active": {
						"type": "boolean"
					},
					"exclude_from_leftover": {
						"type": "boolean"
					},
					"pay_off_monthly": {
						"type": "boolean"
					},
					"created_at": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"type",
					"balance",
					"is_active",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreatePaymentSourceRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/PaymentSourceType"
					},
					"balance": {
						"type": "number",
						"format": "double"
					},
					"is_active": {
						"type": "boolean"
					},
					"exclude_from_leftover": {
						"type": "boolean"
					},
					"pay_off_monthly": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"type",
					"balance"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdatePaymentSourceRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/PaymentSourceType"
					},
					"balance": {
						"type": "number",
						"format": "double"
					},
					"is_active": {
						"type": "boolean"
					},
					"exclude_from_leftover": {
						"type": "boolean"
					},
					"pay_off_monthly": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"BillInstance": {
				"properties": {
					"id": {
						"type": "string"
					},
					"bill_id": {
						"type": "string",
						"nullable": true
					},
					"month": {
						"type": "string"
					},
					"billing_period": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"expected_amount": {
						"type": "number",
						"format": "double"
					},
					"actual_amount": {
						"type": "number",
						"format": "double"
					},
					"payments": {
						"items": {
							"$ref": "#/components/schemas/Payment"
						},
						"type": "array"
					},
					"occurrences": {
						"items": {
							"$ref": "#/components/schemas/Occurrence"
						},
						"type": "array"
					},
					"is_default": {
						"type": "boolean"
					},
					"is_paid": {
						"type": "boolean"
					},
					"is_closed": {
						"type": "boolean"
					},
					"is_adhoc": {
						"type": "boolean"
					},
					"is_payoff_bill": {
						"type": "boolean"
					},
					"payoff_source_id": {
						"type": "string"
					},
					"due_date": {
						"type": "string"
					},
					"closed_date": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"category_id": {
						"type": "string"
					},
					"payment_source_id": {
						"type": "string"
					},
					"created_at": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"bill_id",
					"month",
					"billing_period",
					"amount",
					"expected_amount",
					"payments",
					"occurrences",
					"is_default",
					"is_paid",
					"is_closed",
					"is_adhoc",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IncomeInstance": {
				"properties": {
					"id": {
						"type": "string"
					},
					"income_id": {
						"type": "string",
						"nullable": true
					},
					"month": {
						"type": "string"
					},
					"billing_period": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"expected_amount": {
						"type": "number",
						"format": "double"
					},
					"actual_amount": {
						"type": "number",
						"format": "double"
					},
					"payments": {
						"items": {
							"$ref": "#/components/schemas/Payment"
						},
						"type": "array"
					},
					"occurrences": {
						"items": {
							"$ref": "#/components/schemas/Occurrence"
						},
						"type": "array"
					},
					"is_default": {
						"type": "boolean"
					},
					"is_paid": {
						"type": "boolean"
					},
					"is_closed": {
						"type": "boolean"
					},
					"is_adhoc": {
						"type": "boolean"
					},
					"due_date": {
						"type": "string"
					},
					"closed_date": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"category_id": {
						"type": "string"
					},
					"payment_source_id": {
						"type": "string"
					},
					"created_at": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"income_id",
					"month",
					"billing_period",
					"amount",
					"expected_amount",
					"payments",
					"occurrences",
					"is_default",
					"is_paid",
					"is_closed",
					"is_adhoc",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VariableExpense": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"payment_source_id": {
						"type": "string"
					},
					"month": {
						"type": "string"
					},
					"created_at": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"amount",
					"payment_source_id",
					"month",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FreeFlowingExpense": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"payment_source_id": {
						"type": "string"
					},
					"month": {
						"type": "string"
					},
					"created_at": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"amount",
					"payment_source_id",
					"month",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.number_": {
				"properties": {},
				"additionalProperties": {
					"type": "number",
					"format": "double"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"MonthlyData": {
				"properties": {
					"month": {
						"type": "string"
					},
					"bill_instances": {
						"items": {
							"$ref": "#/components/schemas/BillInstance"
						},
						"type": "array"
					},
					"income_instances": {
						"items": {
							"$ref": "#/components/schemas/IncomeInstance"
						},
						"type": "array"
					},
					"variable_expenses": {
						"items": {
							"$ref": "#/components/schemas/VariableExpense"
						},
						"type": "array"
					},
					"free_flowing_expenses": {
						"items": {
							"$ref": "#/components/schemas/FreeFlowingExpense"
						},
						"type": "array"
					},
					"bank_balances": {
						"$ref": "#/components/schemas/Record_string.number_"
					},
					"is_read_only": {
						"type": "boolean"
					},
					"created_at": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					}
				},
				"required": [
					"month",
					"bill_instances",
					"income_instances",
					"variable_expenses",
					"free_flowing_expenses",
					"bank_balances",
					"is_read_only",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CategoryType": {
				"type": "string",
				"enum": [
					"bill",
					"income",
					"variable"
				]
			},
...
			"CategorySection": {
				"properties": {
					"category": {
						"properties": {
							"type": {
								"$ref": "#/components/schemas/CategoryType"
							},
							"sort_order": {
								"type": "number",
								"format": "double"
							},
							"color": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"sort_order",
							"color",
							"name",
							"id"
						],
						"type": "object"
					},
					"items": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/BillInstanceDetailed"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/IncomeInstanceDetailed"
								},
								"type": "array"
							}
						]
					},
					"subtotal": {
						"properties": {
							"actual": {
								"type": "number",
								"format": "double"
							},
							"expected": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"actual",
							"expected"
						],
						"type": "object"
					}
				},
				"required": [
					"category",
					"items",
					"subtotal"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SectionTally": {
				"properties": {
					"expected": {
						"type": "number",
						"format": "double"
					},
					"actual": {
						"type": "number",
						"format": "double"
					},
					"remaining": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"expected",
					"actual",
					"remaining"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LeftoverBreakdown": {
				"properties": {
					"bankBalances": {
						"type": "number",
						"format": "double"
					},
					"remainingIncome": {
						"type": "number",
						"format": "double"
					},
					"remainingExpenses": {
						"type": "number",
						"format": "double"
					},
					"leftover": {
						"type": "number",
						"format": "double"
					},
					"isValid": {
						"type": "boolean"
					},
					"hasActuals": {
						"type": "boolean"
					},
					"missingBalances": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"errorMessage": {
						"type": "string"
					}
				},
				"required": [
					"bankBalances",
					"remainingIncome",
					"remainingExpenses",
					"leftover",
					"isValid"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PayoffSummary": {
				"properties": {
					"paymentSourceId": {
						"type": "string"
					},
					"paymentSourceName": {
						"type": "string"
					},
					"balance": {
						"type": "number",
						"format": "double"
					},
					"paid": {
						"type": "number",
						"format": "double"
					},
					"remaining": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"paymentSourceId",
					"paymentSourceName",
					"balance",
					"paid",
					"remaining"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DetailedMonthResponse": {
				"properties": {
					"month": {
						"type": "string"
					},
					"billSections": {
						"items": {
							"$ref": "#/components/schemas/CategorySection"
						},
						"type": "array"
					},
					"incomeSections": {
						"items": {
							"$ref": "#/components/schemas/CategorySection"
						},
						"type": "array"
					},
					"tallies": {
						"properties": {
							"totalIncome": {
								"$ref": "#/components/schemas/SectionTally"
							},
							"adhocIncome": {
								"$ref": "#/components/schemas/SectionTally"
							},
							"income": {
								"$ref": "#/components/schemas/SectionTally"
							},
							"totalExpenses": {
								"$ref": "#/components/schemas/SectionTally"
							},
							"ccPayoffs": {
								"$ref": "#/components/schemas/SectionTally"
							},
							"adhocBills": {
								"$ref": "#/components/schemas/SectionTally"
							},
							"bills": {
								"$ref": "#/components/schemas/SectionTally"
							}
						},
						"required": [
							"totalIncome",
							"adhocIncome",
							"income",
							"totalExpenses",
							"ccPayoffs",
							"adhocBills",
							"bills"
						],
						"type": "object"
					},
					"leftover": {
						"type": "number",
						"format": "double"
					},
					"leftoverBreakdown": {
						"$ref": "#/components/schemas/LeftoverBreakdown"
					},
					"payoffSummaries": {
						"items": {
							"$ref": "#/components/schemas/PayoffSummary"
						},
						"type": "array"
					},
					"bankBalances": {
						"$ref": "#/components/schemas/Record_string.number_"
					},
					"lastUpdated": {
						"type": "string"
					}
				},
				"required": [
					"month",
					"billSections",
					"incomeSections",
					"tallies",
					"leftover",
					"leftoverBreakdown",
					"payoffSummaries",
					"bankBalances",
					"lastUpdated"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MonthSummary": {
				"properties": {
					"month": {
						"type": "string"
					},
					"totalIncome": {
						"type": "number",
						"format": "double"
					},
					"totalExpenses": {
						"type": "number",
						"format": "double"
					},
					"leftover": {
						"type": "number",
						"format": "double"
					},
					"isLocked": {
						"type": "boolean"
					}
				},
				"required": [
					"month",
					"totalIncome",
					"totalExpenses",
					"leftover",
					"isLocked"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MonthExistsResponse": {
				"properties": {
					"exists": {
						"type": "boolean"
					},
					"isLocked": {
						"type": "boolean"
					}
				},
				"required": [
					"exists",
					"isLocked"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GenerateMonthRequest": {
				"properties": {
					"force": {
						"type": "boolean",
						"description": "Force regeneration even if month exists"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"LockMonthRequest": {
				"properties": {
					"lock": {
						"type": "boolean",
						"description": "Lock (true) or unlock (false) the month"
					}
				},
				"required": [
					"lock"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateBankBalancesRequest": {
				"properties": {
					"balances": {
						"$ref": "#/components/schemas/Record_string.number_",
						"description": "Map of payment source ID to balance in cents"
					}
				},
				"required": [
					"balances"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "BudgetForFun API",
		"version": "1.0.0",
		"description": "BudgetForFun API - Personal budget tracking application",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "BudgetForFun Team"
		}
	},
	"paths": {
...
		"/api/payment-sources": {
			"get": {
				"operationId": "GetPaymentSources",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PaymentSource"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": "ps-001",
												"name": "Checking Account",
												"type": "bank_account",
												"balance": 250000,
												"is_active": true,
												"exclude_from_leftover": false,
												"pay_off_monthly": false,
												"created_at": "2025-01-01T00:00:00Z",
												"updated_at": "2025-01-01T00:00:00Z"
											}
										]
									}
								}
							}
						}
					}
				},
				"description": "Get all payment sources",
				"summary": "List all payment sources",
				"tags": [
					"Payment Sources"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreatePaymentSource",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentSource"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Create a new payment source",
				"summary": "Create a payment source",
				"tags": [
					"Payment Sources"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePaymentSourceRequest"
							}
						}
					}
				}
			}
		},
		"/api/payment-sources/{id}": {
			"put": {
				"operationId": "UpdatePaymentSource",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentSource"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Update an existing payment source",
				"summary": "Update a payment source",
				"tags": [
					"Payment Sources"
				],
				"security": [],
				"parameters": [
					{
						"description": "The payment source ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePaymentSourceRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletePaymentSource",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Delete a payment source",
				"summary": "Delete a payment source",
				"tags": [
					"Payment Sources"
				],
				"security": [],
				"parameters": [
					{
						"description": "The payment source ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/months/{month}": {
			"get": {
				"operationId": "GetMonth",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MonthlyData"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Get monthly data for a specific month",
				"summary": "Get month data",
				"tags": [
					"Months"
				],
				"security": [],
				"parameters": [
					{
						"description": "Month in YYYY-MM format",
						"in": "path",
						"name": "month",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteMonth",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Delete a month",
				"summary": "Delete month",
				"tags": [
					"Months"
				],
				"security": [],
				"parameters": [
					{
						"description": "Month in YYYY-MM format",
						"in": "path",
						"name": "month",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/months/detailed/{month}": {
			"get": {
				"operationId": "GetDetailedMonth",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DetailedMonthResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Get detailed view for a month (used by the main UI)",
				"summary": "Get detailed month view",
				"tags": [
					"Months"
				],
				"security": [],
				"parameters": [
					{
						"description": "Month in YYYY-MM format",
						"in": "path",
						"name": "month",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/months/summary/{month}": {
			"get": {
				"operationId": "GetMonthSummary",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MonthSummary"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Get month summary (income, expenses, leftover)",
				"summary": "Get month summary",
				"tags": [
					"Months"
				],
				"security": [],
				"parameters": [
					{
						"description": "Month in YYYY-MM format",
						"in": "path",
						"name": "month",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/months/exists/{month}": {
			"get": {
				"operationId": "MonthExists",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MonthExistsResponse"
								}
							}
						}
					}
				},
				"description": "Check if a month exists",
				"summary": "Check month exists",
				"tags": [
					"Months"
				],
				"security": [],
				"parameters": [
					{
						"description": "Month in YYYY-MM format",
						"in": "path",
						"name": "month",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/months/generate/{month}": {
			"post": {
				"operationId": "GenerateMonth",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MonthlyData"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Generate a new month from bill/income templates",
				"summary": "Generate month",
				"tags": [
					"Months"
				],
				"security": [],
				"parameters": [
					{
						"description": "Month in YYYY-MM format",
						"in": "path",
						"name": "month",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GenerateMonthRequest"
							}
						}
					}
				}
			}
		},
		"/api/months/create/{month}": {
			"post": {
				"operationId": "CreateMonth",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MonthlyData"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Create a month (same as generate)",
				"summary": "Create month",
				"tags": [
					"Months"
				],
				"security": [],
				"parameters": [
					{
						"description": "Month in YYYY-MM format",
						"in": "path",
						"name": "month",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/months/sync/{month}": {
			"post": {
				"operationId": "SyncMonth",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MonthlyData"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Sync month with current bill/income templates",
				"summary": "Sync month",
				"tags": [
					"Months"
				],
				"security": [],
				"parameters": [
					{
						"description": "Month in YYYY-MM format",
						"in": "path",
						"name": "month",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/months/lock/{month}": {
			"post": {
				"operationId": "LockMonth",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MonthlyData"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Lock or unlock a month",
				"summary": "Lock/unlock month",
				"tags": [
					"Months"
				],
				"security": [],
				"parameters": [
					{
						"description": "Month in YYYY-MM format",
						"in": "path",
						"name": "month",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LockMonthRequest"
							}
						}
					}
				}
			}
		},
		"/api/months/bank-balances/{month}": {
			"put": {
				"operationId": "UpdateBankBalances",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MonthlyData"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"description": "Update bank balances for a month",
				"summary": "Update bank balances",
				"tags": [
					"Months"
				],
				"security": [],
				"parameters": [
					{
						"description": "Month in YYYY-MM format",
						"in": "path",
						"name": "month",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateBankBalancesRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}